+++ egui_web/CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to the `egui_web` integration will be noted in this file.
 * Shift-scroll will now result in horizontal scrolling ([#1136](https://github.com/emilk/egui/pull/1136)).
 * Updated `epi::IntegrationInfo::web_location_hash` on `hashchange` event ([#1140](https://github.com/emilk/egui/pull/1140)).
 * Panics will now be logged using `console.error`.
+* Parse and percent-decode the web location query string ([#1258](https://github.com/emilk/egui/pull/1258)).
 
 
 ## 0.16.0 - 2021-12-29